[miner] Second pass
