+++ src/Chainweb/Version.hs
@@ -40,6 +40,8 @@ module Chainweb.Version
 , window
 -- ** Date-based Transaction Disabling
 , txSilenceEndDate
+-- ** Testing
+, testMiners
 
 -- * Typelevel ChainwebVersion
 , ChainwebVersionT(..)
@@ -573,3 +575,7 @@ txSilenceEndDate FastTimedCPM{} = Nothing
 txSilenceEndDate Development = Nothing
 txSilenceEndDate Testnet02 = Nothing
 txSilenceEndDate Mainnet01 = Just [timeMicrosQQ| 2019-12-05T00:00:00.0 |]
+
+-- | O(n).
+testMiners :: ChainwebVersion -> Natural
+testMiners = int . HS.size . chainIds
